refactor(jina): 컴포넌트 추출 및 타입 추출 #14
