+++ src/shared/ui/NavBar/PlayList/styled.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+import { CoverWrapper } from '../styled';
+
+export const PlayListInfo = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 30px;
+`;
+
+export const PlayListCoverWrapper = styled(CoverWrapper)``;
