+++ src/shared/ui/NavBar/constants/index.ts
@@ -2,7 +2,7 @@ import IconBoardActiveSrc from 'shared/assets/image/icons/nav-bar/icon-board-act
 import IconBoard from 'shared/assets/image/icons/nav-bar/icon-board.png';
 import IconHomeActiveSrc from 'shared/assets/image/icons/nav-bar/icon-home-active.png';
 import IconHomeSrc from 'shared/assets/image/icons/nav-bar/icon-home.png';
-import { NavItem } from 'shared/types/navbar';
+import { NavItem } from '../type';
 
 export const NAV_ITEM: NavItem[] = [
   {
