+++ src/shared/ui/NavBar/type.ts
@@ -1,3 +1,9 @@
+type NavBarSize = 'small' | 'large';
+
+export interface NavBarSizeProps {
+  size: NavBarSize;
+}
+
 export type NavItem = {
   text: string;
   href: string;
